manual-instrumentation/userbackend-flask/pyproject.toml
[tool.poetry]
name = "userbackend-flask"
version = "0.1.0"
description = "Flask user backend with OpenTelemetry manual instrumentation"
authors = ["LFS148"]

[tool.poetry.dependencies]
python = "^3.12"
faker = "28.4.1"
flask = "3.0.3"
opentelemetry-api = "1.39.0"
opentelemetry-sdk = "1.39.0"
opentelemetry-exporter-otlp = "1.39.0"
opentelemetry-exporter-prometheus = "0.60b0"
prometheus-client = "0.24.1"
psutil = "7.2.2"
requests = "2.32.3"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
